+++ src/App.js
@@ -1,26 +1,42 @@
-import React from "react";
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Navbar from "./components/navbar/Navbar.jsx";
+import { React, useState, Component } from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Men from "./components/pages/Men.jsx";
 import Contact from "./components/pages/Contact.jsx";
 import SignUp from "./components/pages/SignUp";
-import Women from "./components/pages/Women.jsx"
-import Cart from "./components/cart/Cart";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import React, { Component } from "react";
+import Women from "./components/pages/Women.jsx";
 import Card from "./components/card/Card";
 import products from "./datasource.json";
 import Wrapper from "./components/wrapper/Wrapper";
-
+import Footer from "./components/footer/Footer.jsx";
+import Header from "./components/Hero/Header.jsx";
+import Navbar from "./components/navbar/Navbar";
+import Cart from "./components/cart/Cart.jsx";
 
 const App = () => {
-  return( 
-<>
-      <Router>
-        
-          <Navbar />
+  const [isShowCart, setIsShowCart] = useState(false);
+  const [cart, setCart] = useState([]);
+
+  const handleAddToCart = (product) => {
+    setCart((prev) => {
+      const findProductInCart = prev.find((item) => item.id === product.id);
+
+      if (findProductInCart) {
+        return prev.map((item) =>
+          item.id === product.id ? { ...item, amount: item.amount + 1 } : item
+        );
+      }
+
+      return [...prev, { ...product, amount: 1 }];
+    });
+  };
+  return (
+    <>
+      <section className="bg-[#333]">
+        <Router>
+          <Navbar cart={cart} setIsShowCart={setIsShowCart} />
+          <Header />
           {/* Wrap Route elements in a Routes component */}
-          <Routes  >
+          <Routes>
             {/* Define routes using the Route component to render different page components at different paths */}
             {/* Define a default route that will render the Home component */}
             <Route path="/women" element={<Women />} />
@@ -30,37 +46,39 @@ const App = () => {
             <Route path="/contact" element={<Contact />} />
             {/* Define a route that will have descendant routes */}
             <Route path="/signup" element={<SignUp />} />
+            <Route
+              path="/"
+              element={
+                <Wrapper>
+                  {products.map((product) => (
+                    <Card
+                      handleAddToCart={handleAddToCart}
+                      key={product.id}
+                      product={product}
+                      name={product.name}
+                      image={product.image}
+                      hoverImage={product.hoverImage}
+                      sizes={product.sizes}
+                      category={product.category}
+                      tags={product.tags}
+                      price={product.price}
+                      location={product.location}
+                    />
+                  ))}
+                  {isShowCart && (
+                    <Cart
+                      cart={cart}
+                      handleAddToCart={handleAddToCart}
+                      setIsShowCart={setIsShowCart}
+                    />
+                  )}
+                </Wrapper>
+              }
+            />
           </Routes>
-       </Router>
-    <section className="h-screen bg-gray-100">
-      <Cart />
-      <Router>
-        <Routes>
-          <Route path="/cart" />
-          <Route
-            path="/"
-            element={
-              <Wrapper>
-                {products.map((product) => (
-                  <Card
-                    key={product.id}
-                    product={product}
-                    name={product.name}
-                    image={product.image}
-                    hoverImage={product.hoverImage}
-                    sizes={product.sizes}
-                    category={product.category}
-                    tags={product.tags}
-                    price={product.price}
-                    location={product.location}
-                  />
-                ))}
-              </Wrapper>
-            }
-          />
-        </Routes>
-      </Router>
-    </section>
+        </Router>
+        <Footer />
+      </section>
     </>
   );
 };
